+++ model/src/main/java/exceptions/pojo/Admin.java
@@ -1,4 +1,4 @@
-package my.pojo;
+package exceptions.pojo;
 
 import lombok.*;
