+++ model/src/main/java/exceptions/pojo/Teacher.java
@@ -1,27 +1,32 @@
-package my.pojo;
+package exceptions.pojo;
 
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import lombok.*;
 
+import java.util.ArrayList;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 
-@EqualsAndHashCode(callSuper = true, exclude = "groups")
-@ToString(callSuper = true, exclude = "groups")
+@EqualsAndHashCode(callSuper = true)
+@ToString(callSuper = true, exclude = {"group", "students"})
+@Data
 @AllArgsConstructor
 @NoArgsConstructor
-@Data
 @With
-public class Student extends AbstractEntity {
+public class Teacher extends AbstractEntity {
+    private List<Salary> salaries = new ArrayList<>();
     private String login;
     private String password;
     private String first_name;
     private String last_name;
     @JsonManagedReference
-    private Set<Group> groups = new HashSet<>();
-    private Integer age;
+    private Group group;
 
-    public Student(int id, String login, String password, String first_name, String last_name, int age) {
+    private Set<Student> students = new HashSet<>();
+    private int age;
+
+    public Teacher(int id, String login, String password, String first_name, String last_name, int age) {
         super(id);
         this.login = login;
         this.password = password;
@@ -31,13 +36,8 @@ public class Student extends AbstractEntity {
     }
 
     @Override
-    public Student withId(Integer id) {
+    public Teacher withId(Integer id) {
         setId(id);
-        return this;
-    }
-
-    public Student withGroup(Group group) {
-        groups.add(group);
         return this;
     }
 }
