+++ repository/src/main/java/all/repos/AbstractRepositoryInMemory.java
@@ -1,6 +1,6 @@
-package my.repos;
+package all.repos;
 
-import my.pojo.AbstractEntity;
+import exceptions.pojo.AbstractEntity;
 
 import java.util.HashSet;
 import java.util.Map;
