+++ repository/src/main/java/exceptions/ApplicationException.java
@@ -1,0 +1,15 @@
+package exceptions;
+
+public class ApplicationException extends RuntimeException{
+    public ApplicationException(String message) {
+        super(message);
+    }
+
+    public ApplicationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public ApplicationException(Throwable cause) {
+        super(cause);
+    }
+}
