+++ repository/src/main/java/exceptions/DatabaseException.java
@@ -1,0 +1,15 @@
+package exceptions;
+
+public class DatabaseException extends ApplicationException{
+    public DatabaseException(String message) {
+        super(message);
+    }
+
+    public DatabaseException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public DatabaseException(Throwable cause) {
+        super(cause);
+    }
+}
